Allow specifying the source location of `#expect()` and `#require()`.
This PR adds an additional argument to the `#expect()` and `#require()` macros that, if specified, replaces the source location derived at macro expansion time and passed to `__check()`. This change is ABI-safe for anyone who has already adopted swift-testing because it does not affect the emitted code or symbols. It is source-safe.

In the typical use case, the existing/old code path is taken at compile time and this PR has no effect. If a developer wants to specify a precise source location, they can do so.

This PR also changes the default behaviour of `#expect()` to generate an instance of `SourceLocation` using default arguments rather than by asking the macro context to provide all info. This is a little bit faster and is more accurate when a freestanding macro expression is embedded in another macro expression, such as:

```swift
  #expect(try foo())
}
```

> [!NOTE]
> The arguments to `SourceLocation.init()` all currently have default values. We might want to change that in the future so that a developer cannot accidentally underspecify an instance. Where a default value is truly required, we can offer a `.here()` factory function that explicitly indicates intent.

Resolves rdar://112974442.
Resolves #15.