This change allows a developer to specify an argument to a test that is the result of a throwing (or asynchronous throwing) function/operation. For example:
```swift
@Test(arguments: try readMenu())
func eatFood(_ food: Food) {
  ...
}
```

These arguments are resolved at runtime during the test planning stage. If an argument collection throws an error while being resolved, it is caught and stored, and the test function's body is replaced with one that records the error as an issue.

There isn't a static mechanism by which we can tell that a test function calls `withKnownIssue {}`. I'm taking the convenient position that, because the issue occurred outside the test function's body, any `withKnownIssue {}` calls in the test function don't apply.