Add a protocol defining custom descriptions for types during expectation failure.
This PR introduces a new protocol, `CustomExpectationFailureRepresentable`, to which types can conform in order to control their representations during an expectation failure. The protocol contains a single property of type `String`.

A default implementation is provided for `Optional` so that `Optional(x)` is not logged in favour of simply `x`. Developers can provide additional implementations for their own types.

I could use feedback on the name of the protocol and its property, in particular!