Add `borrowing` to more places that pass `Event` and `Event.Context`.
This PR reduces the number of copies of `Event` and `Event.Context` that may be made when an event occurs by marking more parameters with these types as `borrowing`.

I also took a moment to update the documentation for some of the event types to explain where their associated values ended up as it may not be clear from initial inspection.