Fix the 256-color terminal check so it doesn't require `_StringProcessing`. (#60)
Use `strstr()` so that we can detect 256-color terminal support on older macOS versions or with older toolchains that don't have the `_StringProcessing` module.

I did _not_ change another use of `contains()` in ExpectationChecking+Macro.swift because it may be sensitive to Unicode nuances whereas a search for the ASCII string ``"256"` isn't.

I also took the opportunity to replace one other usage of `#available(explicit_versions_here)` with an availability define. I sorted the availability defines in the package manifest too.