Rethink tag colors as secondary package data.
This PR migrates tag colors from being specific instances of `Tag` to being out-of-band data contained outside of a package. We define a new folder in the user's home folder named `".swift-testing"` that contains out-of-band data specific to that user. In particular, we store tag colors in a file named `"tag-colors.json"` as a JSON object (dictionary.)

Note that we use `JSONDecoder` from Foundation to implement decoding. Although we do not want to add dependencies to Foundation (as it could prevent that package from someday adopting swift-testing), this code is contained in `XCTestScaffold` which already uses XCTest and, indirectly, Foundation, so this does not constitute a new dependency.

This path does not necessarily represent a final home for this data. We may instead want to put the file in the package folder or manifest, or in some location important to Xcode or other IDEs.