Add an _ersatz_ filtering mechanism to `XCTestScaffold`. (#77)
* Add an _ersatz_ filtering mechanism to `XCTestScaffold`.

Because SwiftPM is not (yet) aware of swift-testing, `swift test --filter` does not have any effect on swift-testing tests. This PR adds temporary support for an environment variable, `"SWT_SELECTED_TEST_IDS"`, to stand in for `--filter` until such time as SwiftPM can be modified to support swift-testing.

Note that the supported value of this environment variable is (intentionally) constrained to just test IDs right now. Matching display names or other metadata about tests is certainly possible, but it's beyond the scope of today's PR.