Support setting a debugger breakpoint to pause when a test fails (#63)
* Support setting a debugger breakpoint to pause when a test fails

* Refine the C breakpoint symbol's declaration

* Update approach to no longer use @_cdecl

While validating this, I discovered @_cdecl creates a C thunk to call the Swift function, instead of giving the actual Swift function the specified C declarataion name, and Swift callers of the function do not call through the thunk and instead directly call the Swift function. This means that the C function thunk cannot be used to create the breakpoint since it will never be called by the testing library.

My new approach is to drop the use of C declarations/symbols altogether and allow setting the breakpoint on the Swift function, and I enhanced the DocC discussion to include a recommended LLDB command for doing this.