Remove a dependency on XCTest. (#98)
* Remove a dependency on XCTest.

This PR removes a static dependency on XCTest in swift-testing. The dependency exists because a single Swift function cannot be _both_ an XCTest test method _and_ a swift-testing test function, so we have a runtime check in our generated thunk that records an issue if a developer manages to produce such code.

This change removes a reference to `XCTestCase` on non-Darwin platforms. Instead, on those platforms we dynamically look up the `XCTest.XCTestCase` class by name at runtime (similar to how we find it on Darwin.)

The generic function that performs this check is now generic over `AnyObject` rather than `XCTestCase`. There is a nominal runtime cost to taking this code path for all reference types, however it is not expected to be a problem except for pathologically deep class hierarchies (which we'd expect would suffer performance issues elsewhere in Swift/the runtime anyway.)