Support: fix `Environment.variable(named:)` on Windows
The return value being `0` may or may not indicate.  We need to
additionally check `GetLastError` to see if `ERROR_ENVVAR_NOT_FOUND` is
set to indicate the missing variable.  Otherwise, the variable is just
an empty string.  This repairs the 2 failing tests on Windows.

Thanks to @grynspan for the fix!