Simplify how we filter out hidden tests. (#115)
Right now, we're checking if tests are hidden in multiple places during execution. This is somewhat inefficient, but it's also error-prone because it means that we can create new code paths that _don't_ correctly check for hidden tests.

This PR consolidates the checks for hidden tests to `Configuration.testFilter` itself. With this change, hidden tests are _always_ filtered out by that filtering function regardless of what value it's set to. The only escape hatch—which we need for our own unit tests only—is to explicitly call the internal-only `Configuration.setTestFilter(toMatch:includeHiddenTests:)` and pass `true`.