Remove the `_TestCases` and `Cases` types.
This PR removes a chunk of infrastructure that we use to type-erase instances of `Test.Case.Generator<S>`. What we _really_ want to do is express `any Sequence<Test.Case> & Sendable`, however compiler work is required to make that possible ([96960993](rdar://96960993)).

In lieu of that compiler work, and instead of rolling our own type-erasing box, we can just use `AnyIterator<Test.Case>` for storage. Combine it with `UncheckedSendable` (since we know our sequence really is sendable and the implementation of `AnyIterator` is safely trivial) and we get what we want without needing to build it ourselves.