(Tests only) Conditionalize imports of Foundation and always import TestingInternals as private. (#131)
This PR conditionalizes imports of Foundation and always import TestingInternals as private within our own test target. This is necessary because some environments may not have Foundation available during testing, and because the lack of `private` on `import TestingInternals` leads to ambiguity and a build error when our targets are statically linked together.