Remove `CartesianProduct` as a distinct type.
This PR removes `CartesianProduct` as a distinct sequence type and instead simply uses its underlying algorithm (`flatMap * map`) directly in the one place where we need it.

Since we no longer need to track the `underestimatedCount` of a sequence of test cases, we don't lose anything here except a few hundred lines of overhead. :D