Rename the SwiftPM entry point.
This PR renames the SwiftPM entry point function such that it does not need to be SPI. By marking it SPI, we unintentionally made it very difficult to use with a toolchain that does not include private module information.

The PR continues to export the old symbol name as used by SwiftPM; once we've updated SwiftPM to match, we will remove the old function.