Avoid constructing a `Test` for an unavailable parameterized test function. (#146)
* Revert "Do not evaluate unavailable test function parameters. (#133)"

This reverts commit 9e105df9549fb7c86d1ee5bdc9128e75bff25efd.

* Avoid constructing a `Test` for an unavailable parameterized test function.

This PR rethinks the work in https://github.com/apple/swift-testing/pull/133. Instead of lazily evaluating individual arguments, we can instead lazily evaluate the entire test. In the event the test is unavailable, we substitute a zero-argument no-op function so that there is no opportunity for us to touch unavailable type metadata. 🤞🏻

Resolves rdar://118996437.