Refine `makeTestFilter()` further.
This PR refines the previous PR for this issue, #153. It removes the internal overload of `makeTestFilter(matching:uncheckedTestFilter:)` because it is redundant: since the result needs to be passed to `uncheckedTestFilter` to preserve its lack of `isHidden`-checking, it doesn't _also_ need to impose its own check.

I did add an internal overload of `makeTestFilter()` to the test target that just returns `selection.contains` as a convenience.

Resolves rdar://119205417 (again.)