Mark `XCTestScaffold` to-be-deprecated.
Currently, the documentation for `XCTestScaffold` has a "warning" callout to indicate that it is a temporary interface. This PR replaces the callout with a formal deprecation attribute so that the type and its members show as deprecated in our documentation. They are marked as deprecated in Swift 100000.0, which is a "magic" version number used in Apple's SDK to indicate "to-be-deprecated" status.