Parallelize tests by default.
This PR enables parallelization by default when using `swift test` and enables support for the `--no-parallel` flag in `swift test` added in https://github.com/apple/swift-package-manager/pull/7231.

The default value of `Configuration.isParallelizationEnabled` remains `false`. Other adopters are free to enable or disable parallelization by default.

`XCTestScaffold` is not affected. It is a shim used with XCTest in Swift 5.10 only and expects serial execution of tests.