Optionally capture source code representations for all traits.
The current macro expansion implementation for `@Test` and `@Suite` checks for known patterns corresponding to _tags_ (i.e. the `.tags()` trait) and injects additional code to capture the source code of each tag.

This change generalizes that logic by adding a requirement to the `Trait` protocol that the macro expansion calls for _all_ traits. It passes the source code of the trait expression as well as the source code of each argument to the trait (if it's a function call.) The macro expansion thus no longer needs to recognize the `Tag.List` trait type specifically.

The source code representations are captured as autoclosures to avoid evaluating them if not needed. The default implementation of the requirement ignores them, so for a typical trait there is no effect. `Tag.List` implements the requirement and maps the argument source code representations to its tags.

This change also removes the overload of `.tags()` that takes a sequence as it makes expressions like `.tags(foo)` ambiguous.