Remove `public` keyword from most swift-syntax imports in the macro target.
This PR removes the explicit `public` keyword from imports of SwiftSyntax and SwiftSyntaxMacros in all files except those that actually declare macro implementations. If compiling using Swift 5.10 or earlier, the `public` keyword is preserved because earlier compilers don't correctly handle a scenario where a module is imported publicly in one file and internally in another.

This change removes a number of warnings of the form:

```
.../swift-testing/Sources/TestingMacros/Support/AttributeDiscovery.swift:11:8: warning: public import of 'SwiftSyntax' was not used in public declarations or inlinable code
public import SwiftSyntax
~~~~~~~^
.../swift-testing/Sources/TestingMacros/Support/AttributeDiscovery.swift:12:8: warning: public import of 'SwiftSyntaxMacros' was not used in public declarations or inlinable code
public import SwiftSyntaxMacros
~~~~~~~^
```

This change also tweaks the declarations of `ExpectMacro` and `RequireMacro` as they indirectly conform to `ExpressionMacro` in a way that may confuse some versions of the Swift compiler.