Capture discrete subexpressions in a `SourceCode` instance.
The `SourceCode` type currently represents the source code, as written by a developer, of some expression such as the body of a `#expect()` macro invocation. This PR enhances `SourceCode` to let it capture, at runtime, information about its various subexpressions and to propagate that information to an expectation or expectation snapshot.

This change allows us to (in a future PR) provide more verbose descriptions of expectation failures when invoking `swift test --verbose`.

Resolves rdar://121382419.