Do not use `String.init(validatingUTF8: [CChar])`.
The overload of `String.init(validatingUTF8:)` that takes an array of `CChar` values rather than a pointer is not really meant to be used. The overload exists for the benefit of the type checker so that future compilers can warn when a developer misuses the implicit array-to-pointer conversion.

This PR changes our wrapper around `sysctlbyname()` to avoid producing a temporary `[CChar]`, thus dodging the problem.

See also https://github.com/apple/swift/pull/71224.

Resolves rdar://122198850.