Implement a test iteration feature.
This PR introduces a test iteration feature similar to the one supported by XCTest. Developers setting up an instance of `Configuration` can specify how many times to repeat a test plan during a test run as well as under what conditions iteration should continue or stop.

A new property, `iterationPolicy`, is added to `Configuration`, with two properties of its own as described above and a couple of convenience factory methods. When a `Runner` is run, its configuration's iteration policy is consulted and the entire test run is repeated as needed. New event kinds have been added for iteration start and end.

SwiftPM integration to come in a later PR in that repository.

Resolves rdar://88665403.