Expand on `Test.current` and why it is optional. (#216)
The current documentation for
[`Test.current`](https://swiftpackageindex.com/apple/swift-testing/main/documentation/testing/test/current)
doesn't explain why it's an optional property. If no test is running, or
if the current thread is detached from the task that's running a test,
then the value is `nil`. Our documentation should reflect that.

(Same for `Test.Case.current`.)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.