Add a `.serial` trait that forces a test to run its content serially.
As of right now, when parallelization is enabled (the default for `swift test`, opt-in for `XCTestScaffold`), all tests scramble to run at the same time. This is generally overall a good thing for performance, because the total time spent running tests is measurably lower than if they ran serially.

However, some tests, such as those that touch shared global state, really need to run serially. This PR introduces a new trait, `.serial` AKA `SerialTrait`, that allows those tests to opt out of parallelization. Note that this is not a global effect: the tests may run in parallel with other unrelated tests, but their test cases will run serially relative to each other, and in the case of suites any contained tests will run serially (and the effect is recursive, so a parameterized test function contained in a suite that is contained in a serialized suite will run its cases serially.)

To avoid teaching the runner about another specialized trait, I wanted to modify the existing `Trait.prepare(for:)` function to take an `inout` `Runner.Plan.Action` value that a trait could then modify. However, `Runner.Plan.Action`'s design is not ready to become API, so instead I have introduced a new protocol `SPIAwareTrait` that refines `Trait` and adds the appropriate requirement.

This change means that `Runner.Plan.Action.run` now has an associated value, because whether or not a test runs parallelized is necessarily part of the step for that test. I've made the associated value an options-y structure so that we can extend it later.

> [!NOTE]
> It's very hard to write unit tests that validate that `.serial` does what it's designed to do; I'm open to writing additional tests if we can think up some that can be implemented in a reliable way.

Resolves rdar://95629402.