Migrate file I/O code to a dedicated `FileHandle` type.
This PR migrates the bulk of our file I/O code into a single file/type,
`FileHandle`. Previously, the code we used to touch the file system and `stderr`
was spread across several Swift files. This has started to become a bit
difficult to maintain, especially when you consider platform-specific
differences (I'm looking at you, Windows…)

So I've moved the code into a single type that wraps a C `FILE *` handle. The
type is move-only and sendable, and abstracts away most of the platform-specific
code we had around files. Access to the underlying file handle is done via a
scoped-access "with" function.