Reduce the sizes of `Event.Kind`, `Issue.Kind`, and `Runner.Plan.Action`. (#219)
Right now, all cases in the `Event.Kind`, `Issue.Kind`, and
`Runner.Plan.Action` enums are direct, i.e. stored inline in instances
of those enums. That requires that the enums be at least as large as
their largest associated values, which turns out to be a fair bit:

| Type | Size (Bytes) |
|-|-|
| `Event` | 464 |
| `Event.Kind` | 345 |
| `Issue` | 225 |
| `Issue.Kind` | 160 |
| `Runner.Plan.Action` | 225 |

Nearly half a kilobyte for an event is excessive. We can reduce the
space needed by storing the associated values of these enums
_indirectly_ (using the `indirect` keyword.) There are cases of both
enums that don't have any associated values, so let's just make _all_
associated values indirect. We end up with:

| Type | Size (Bytes) |
|-|-|
| `Event` | 120 |
| `Event.Kind` | 8 |
| `Issue` | 73 |
| `Issue.Kind` | 8 |
| `Runner.Plan.Action` | 8 |

The tradeoff is that indirect storage is basically a pointer
dereference, so we lose locality—but the frequency with which events and
issues are raised doesn't make the performance cost worth worrying
about.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.