Add the ability for a trait to run code before and after a test (#86)
Add the ability for a trait to run code before and after a test.

### Motivation:

As laid out in the [vision
document](https://github.com/apple/swift-testing/blob/main/Documentation/Vision.md#trait-extensibility),
we want traits to become extensible, and be able to invoke code *before*
and *after* a test's body is invoked.

This PR provides the foundation that will not just give test authors a
lot more power and flexibility in what they can achieve with traits, but
will allow us (in a future PR) to re-implement e.g. the timeout behavior
to be less special-cased.

### Modifications:

I added a new protocol `CustomExecutionTrait` which inherits from
`Trait`, that has a single additional requirement: an `execute` function
that provides a function that eventually executes the test body in its
argument so that trait authors can invoke it in whatever way they need,
ie. have the ability to execute code before and after they invoke that
function.

An example of such a trait is

```swift
struct DoSomethingBeforeAndAfterTrait: CustomExecutionTrait & TestTrait {

  func execute(_ function: () async throws -> Void, for test: Test) async throws {
    something()
    try await function()
    somethingElse()
  }
}
```

(or see `CustomTraitTest`).

### Result:

Test authors now have a much more flexible system to write traits,
allowing them to execute custom code before and after a test body is
run.