Add 'README' to Documentation directory, and remove .docc symlink (#224)
This adds a new `README` file to the `Documentation/` subdirectory with
links to all the documentation resources related to this project. It
also deletes the `Testing.docc` symlink in this directory.

### Motivation:

The main motivation is to delete the symlink because it causes searching
for content in IDEs to incorrectly behave as though each Markdown .md
file is at two locations in the filesystem, resulting in duplicate
matches. But to ensure the .docc bundle is still easy to find, I've
added a README file containing a link to it.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.