Represent type information for test parameters and their runtime arguments (#229)
This adds support for collecting and representing type information for
parameters of test functions which have them, and runtime arguments
passed to them.

### Motivation:

When examining test results, it's not always obvious what the type of an
argument to a parameterized test function was. A string description of
the value is captured and included in results, but this does not always
explicitly indicate a value's type. This PR includes support for
capturing a representation of a type, in both qualified and unqualified
formats, so that tools can display this and provide more detailed and
actionable results.

### Modifications:

- Add a new type `TypeInfo`, currently SPI, which provides information
about a type.
- Expose an instance of `TypeInfo` from `Test.Case.Argument.Snapshot`,
`Test.Parameter`, and `Expression`.
- Enhance the `@Test` macro to include the type of a function parameter.
- Include the qualified representation of each argument's type when
`--verbose` mode is enabled in `HumanReadableOutputRecorder`.
- Add and adjust unit tests.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

Resolves rdar://119614405